Place the rate indicator up top
And to the left, no way to miss it. Before, it was alone in its own row
at the bottom of everything.

Also style it more prominently.

More importantly, it merges the rows for this and the video selector,
going from 4 to 3. It makes the widget more handy when e.g. you're in
fullscreen mode on YouTube.