Add a persistent mode where clicking doesn't remove the controls
Activate with Ctrl+P. (I've disabled Ctrl+P = Print Dialog when the
bookmarklet is active.)

This may only work on YouTube (my use case for all this tweaking), where
you can invoke the bookrmarklet's controls, enter focus mode, and the
controls are still there in the upper right corner.

Before this: the underlying video/page (below the video_speed_cover
element added by the bookmarklet) didn't respond to mouse clicking or
hovering.

The cover element intercepted those events and, on click, dismissed the
playback speed controls.

It did respond to keyboard shortcuts, because those are set directly to
the document, and don't have to go through the cover first.

After this: the underlying video responds to click and hover events
(useful e.g. when you want to see how much of the video still to watch).

It also doesn't dismiss the widget on click (useful e.g. when you want
to skip ahead of the video by a large margin).

My use case for all of this: I'm wathing a Final Fantasy playthrough. I
watch on fullscreen (Focus Mode), I want to speed up boring battles and
then go back to 1x speed (or similar) when the battle ends. With these
changes, I can do all of that without exiting Focus Mode.